style: display emojis horizontally and lock link at bottom
Refactor HTML and CSS to display emojis in a horizontal stack, one on top of each other, filling 80% of the screen height. The link "Inspired by 🐮.ws" is now locked at the bottom of the screen and takes up a maximum of 20% of the screen height. This change improves the layout and readability of the page.

BREAKING CHANGE: The display of the emojis and the position of the link have been changed, affecting the layout of the page. This may require updates to other parts of the application that rely on the previous layout.